Created a new class named Notification and  implementation of notifications when services throw errors

+++ ApiTresCamadas/src/DevIo.Domain/Services/FornecedorService.cs
@@ -1,5 +1,7 @@
 ﻿using DevIo.Domain.Interfaces;
 using DevIo.Domain.Models;
+using DevIo.Domain.Models.Validations;
+using FluentValidation.Validators;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,24 +13,64 @@ namespace DevIo.Domain.Services
     public class FornecedorService : BaseService, IFornecedorService
     {
         private readonly IFornecedorRepository _fornecedorRepository;
-        public FornecedorService(IFornecedorRepository fornecedorRepository)
+        public FornecedorService(IFornecedorRepository fornecedorRepository, 
+            INotificador notificador) : base(notificador)
         {
             _fornecedorRepository = fornecedorRepository;
         }
 
         public async Task Adicionar(Fornecedor fornecedor)
         {
+            if (!ExecutarValidacao(new FornecedorValidation(), fornecedor) ||
+                !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
+
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+            {
+                Notificar("Já existe um fornecedor com este documento informado.");
+                return;
+            }
+
             await _fornecedorRepository.Adicionar(fornecedor);
         }
 
         public async Task Atualizar(Fornecedor fornecedor)
         {
+            if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
+
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+            {
+                Notificar("Já existe um fornecedor com este documento informado.");
+                return;
+            }
+
             await _fornecedorRepository.Atualizar(fornecedor);
         }
 
 
         public async Task Remover(Guid id)
         {
+            var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+            if (fornecedor == null)
+            {
+                Notificar("Fornecedor não existe!");
+                return;
+            }
+
+            if (fornecedor.Produtos.Any())
+            {
+                Notificar("O fornecedor possui produtos.");
+                return;
+            }
+
+            var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+            if (endereco != null)
+            {
+                await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+
+            }
+
             await _fornecedorRepository.Remover(id);
         }
